+++ app/case/jobit/page.tsx
@@ -1,6 +1,5 @@
 import CaseLayout from '@/components/Case/CaseLayout';
-
-import { jobitCase } from '@/utils/cases';
+import { jobitCase } from '@/utils';
 
 const JobitCaseStudy = () => {
   return <CaseLayout data={jobitCase} />;
