+++ app/case/morrent/page.tsx
@@ -1,5 +1,5 @@
 import CaseLayout from '@/components/Case/CaseLayout';
-import { morrentCase } from '@/utils/cases';
+import { morrentCase } from '@/utils';
 
 const MorrentCaseStudy = () => {
   return <CaseLayout data={morrentCase} />;
