+++ app/page.tsx
@@ -1,12 +1,10 @@
-import React from 'react';
-
-import Hero from '@/components/Home/Hero';
-import Skills from '@/components/Home/Skills/Skills';
-import Services from '@/components/Home/Services';
+import GetInTouch from '@/components/GetInTouch';
 import Experience from '@/components/Home/Experience';
 import FeaturedProjects from '@/components/Home/FeaturedProjects';
+import Hero from '@/components/Home/Hero';
 import Reviews from '@/components/Home/Reviews';
-import GetInTouch from '@/components/GetInTouch';
+import Services from '@/components/Home/Services';
+import Skills from '@/components/Home/Skills/Skills';
 
 export default function Home() {
   return (
