+++ components/Contact/ContactHero.tsx
@@ -1,0 +1,12 @@
+const ContactHero = () => {
+  return (
+    <section className='flex flex-col items-center bg-background-2 py-16'>
+      <h1 className='text-secondary'>Get In Touch</h1>
+      <p className='font-light text-white-500 dark:text-white'>
+        {"Let's Collaborate on Your Next Project"}
+      </p>
+    </section>
+  );
+};
+
+export default ContactHero;
