+++ components/Contact/index.ts
@@ -1,0 +1,6 @@
+import ContactContent from './ContactContent';
+import ContactForm from './ContactForm';
+import ContactHero from './ContactHero';
+import ContactInfo from './ContactInfo';
+
+export { ContactContent, ContactForm, ContactHero, ContactInfo };
