+++ components/Home/Skills/index.ts
@@ -1,0 +1,6 @@
+import ProgressBar from './ProgressBar';
+import SkillIcons from './SkillIcons';
+import Skills from './Skills';
+import SkillsTab from './SkillsTab';
+
+export { ProgressBar, SkillIcons, Skills, SkillsTab };
