+++ components/Home/index.ts
@@ -1,0 +1,21 @@
+import CopyButton from './CopyButton';
+import Experience from './Experience';
+import FeaturedProjects from './FeaturedProjects';
+import Hello from './Hello';
+import Hero from './Hero';
+import ReviewButton from './ReviewButton';
+import Reviews from './Reviews';
+import Services from './Services';
+
+export {
+  CopyButton,
+  Experience,
+  FeaturedProjects,
+  Hello,
+  Hero,
+  ReviewButton,
+  Reviews,
+  Services,
+};
+
+export * from './Skills';
