+++ components/Nav/Desktop/index.ts
@@ -1,0 +1,3 @@
+import DesktopNav from './DesktopNav';
+
+export { DesktopNav };
