+++ components/Nav/MobileNav/index.ts
@@ -1,0 +1,4 @@
+import MobileNav from './MobileNav';
+import MobileNavLinks from './MobileNavLinks';
+
+export { MobileNav, MobileNavLinks };
