+++ components/SVG/Hamburger.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+'use client';
+
 import { motion } from 'framer-motion';
 
 const Path = (props: any) => (
