+++ components/SVG/index.ts
@@ -1,0 +1,3 @@
+import Hamburger from './Hamburger';
+
+export { Hamburger };
