+++ components/index.ts
@@ -1,0 +1,4 @@
+export * from './SVG';
+export * from './Contact';
+export * from './Home';
+export * from './Nav';
