+++ types/componentTypes.d.ts
@@ -1,5 +1,12 @@
 import { ReactNode } from 'react';
 
+export interface emailData {
+  name: string;
+  email: string;
+  project: string;
+  contact: string;
+}
+
 export interface SettingsType {
   value: string;
   label: string;
