--- utils/components.ts
@@ -1,6 +1,0 @@
-export interface emailData {
-  name: string;
-  email: string;
-  project: string;
-  contact: string;
-}
