+++ utils/contact/index.ts
@@ -1,0 +1,28 @@
+export const contactQuestions = [
+  {
+    question: "What's your name?",
+    type: 'text',
+    name: 'name',
+    invalid: 'Please enter your Full Name',
+  },
+  {
+    question: "What's your email?",
+    type: 'email',
+    name: 'email',
+    invalid: 'Please enter your Email',
+    mismatch: 'Please enter a valid Email.',
+  },
+  {
+    question: 'Write something about your project goals and timeframe',
+    type: 'text',
+    name: 'project',
+    invalid: 'Please enter information about your project.',
+  },
+  {
+    question: 'How to reach out to you back?',
+    exempli: ' eg. phone number or email',
+    type: 'text',
+    name: 'contact',
+    invalid: 'Let me know how I can reach you.',
+  },
+];
