+++ utils/iconsImages.ts
@@ -1,0 +1,121 @@
+import Jobit from '@/public/projects/jobit.png';
+import Morrent from '@/public/projects/morrent.png';
+import JobitCase from '@/public/cases/projects/jobit.png';
+import MorrentCase from '@/public/cases/projects/morrent.png';
+import GlobeIcon from '@/public/cases/globe.svg';
+import GithubIcon from '@/public/cases/github.svg';
+import RightArrow from '@/public/cases/arrowright.svg';
+import DefIcon from '@/public/cases/def.png';
+import DesignIcon from '@/public/cases/design.png';
+import PlanIcon from '@/public/cases/plan.png';
+import DevIcon from '@/public/cases/dev.png';
+import DeployIcon from '@/public/cases/deploy.png';
+import DefLightIcon from '@/public/cases/defLight.png';
+import DesignLightIcon from '@/public/cases/designLight.png';
+import PlanLightIcon from '@/public/cases/planLight.png';
+import DevLightIcon from '@/public/cases/devLight.png';
+import DeployLightIcon from '@/public/cases/deployLight.png';
+import ChallengeIcon from '@/public/cases/challenge.png';
+import SuccessIcon from '@/public/cases/success.png';
+import TypeScriptIcon from '@/public/cases/tech/ts.svg';
+import NextIcon from '@/public/cases/tech/next.svg';
+import CSSIcon from '@/public/cases/tech/css.svg';
+import ReactIcon from '@/public/cases/tech/react.svg';
+import HTMLIcon from '@/public/cases/tech/html.svg';
+import MotionIcon from '@/public/cases/tech/motion.svg';
+import TailwindIcon from '@/public/cases/tech/tailwind.svg';
+import GithubTechIcon from '@/public/cases/tech/github.svg';
+import GitIcon from '@/public/cases/tech/git.svg';
+import NodeIcon from '@/public/cases/tech/node.svg';
+import SupabaseIcon from 'public/cases/tech/supabase.svg';
+import AllMight from '@/public/people/allmight.jpg';
+import PlaceholderUser from '@/public/people/placeholder-user.jpg';
+import ArrowLeft from '@/public/icons/arrow-left.svg';
+import ArrowRight from '@/public/icons/arrow-right.svg';
+import FullStar from '@/public/icons/fullstar.svg';
+import AnalyticsIcon from '@/public/icons/services/analytics.svg';
+import MonitorIcon from '@/public/icons/services/monitor.svg';
+import ServerIcon from '@/public/icons/services/server.svg';
+import OptimizeIcon from '@/public/icons/services/optimize.svg';
+import InstagramIcon from '@/public/icons/social/instagram.svg';
+import XIcon from '@/public/icons/social/x.svg';
+import LinkedInIcon from '@/public/icons/social/linkedin.svg';
+import GithubSocialIcon from '@/public/icons/social/github.svg';
+import PhoneIcon from '@/public/phone.svg';
+import EmailIcon from '@/public/email.svg';
+import MorrentProblem from '@/public/cases/morrentproblem.png';
+import JSIcon from '@/public/icons/companies/js.svg';
+import TSIcon from '@/public/icons/companies/typescript.svg';
+import FramerIcon from '@/public/icons/companies/framer.svg';
+import MongoIcon from '@/public/icons/companies/mongodb.svg';
+import MaterialIcon from '@/public/icons/companies/materialui.svg';
+import SassIcon from '@/public/icons/companies/sass.svg';
+import HeadlessIcon from '@/public/icons/companies/headless.svg';
+import RadixIcon from '@/public/icons/companies/radix.png';
+import Download from '@/public/icons/download.svg';
+import Divider from '@/public/icons/divider.svg';
+import SunIcon from '@/public/icons/sun.svg';
+import MoonIcon from '@/public/icons/moon.svg';
+import NavMonitorIcon from '@/public/icons/monitor.svg';
+
+export {
+  SunIcon,
+  MoonIcon,
+  NavMonitorIcon,
+  Download,
+  Divider,
+  GlobeIcon,
+  GithubIcon,
+  RightArrow,
+  DefIcon,
+  DesignIcon,
+  PlanIcon,
+  DevIcon,
+  DeployIcon,
+  DefLightIcon,
+  DesignLightIcon,
+  PlanLightIcon,
+  DevLightIcon,
+  DeployLightIcon,
+  ChallengeIcon,
+  SuccessIcon,
+  Jobit,
+  Morrent,
+  TypeScriptIcon,
+  NextIcon,
+  CSSIcon,
+  ReactIcon,
+  HTMLIcon,
+  MotionIcon,
+  TailwindIcon,
+  GithubTechIcon,
+  GitIcon,
+  NodeIcon,
+  SupabaseIcon,
+  ArrowLeft,
+  ArrowRight,
+  FullStar,
+  AnalyticsIcon,
+  MonitorIcon,
+  ServerIcon,
+  OptimizeIcon,
+  InstagramIcon,
+  XIcon,
+  LinkedInIcon,
+  GithubSocialIcon,
+  PhoneIcon,
+  EmailIcon,
+  JobitCase,
+  MorrentCase,
+  AllMight,
+  PlaceholderUser,
+  MorrentProblem,
+  JSIcon,
+  TSIcon,
+  FramerIcon,
+  MaterialIcon,
+  SassIcon,
+  HeadlessIcon,
+  RadixIcon,
+  MongoIcon,
+};
